+++ microservices/payment-stripe/src/constants/charge-refund-status.ts
@@ -1,0 +1,8 @@
+enum ChargeRefundStatus {
+  NO_REFUND = 'noRefund',
+  PARTIAL_REFUND = 'partialRefund',
+  // Full transaction amount refunded to sender. No matter with transfer or not
+  FULL_REFUND = 'fullRefund',
+}
+
+export default ChargeRefundStatus;
