+++ microservices/payment-stripe/src/constants/refund-status.ts
@@ -1,0 +1,10 @@
+enum RefundStatus {
+  INITIAL = 'initial',
+  IN_PROCESS = 'inProcess',
+  REQUIRES_ACTION = 'requiresAction',
+  SUCCESS = 'success',
+  ERROR = 'error',
+  CANCELED = 'canceled',
+}
+
+export default RefundStatus;
