+++ microservices/payment-stripe/src/constants/stripe-refund-status.ts
@@ -1,0 +1,9 @@
+enum StripeRefundStatus {
+  PENDING = 'pending',
+  REQUIRES_ACTION = 'requires_action',
+  SUCCEEDED = 'succeeded',
+  FAILED = 'failed',
+  CANCELED = 'canceled',
+}
+
+export default StripeRefundStatus;
