+++ microservices/payment-stripe/src/services/webhook-handlers/index.ts
@@ -1,5 +1,12 @@
+import { EntityManager } from 'typeorm';
+import Account from './account';
+import ApplicationFee from './application-fee';
 import Charge from './charge';
 import Customer from './customer';
+import PaymentIntent from './payment-intent';
+import PaymentMethod from './payment-method';
+import SetupIntent from './setup-intent';
+import Transfer from './transfer';
 
 /**
  * Webhook handlers
@@ -18,19 +25,54 @@ class WebhookHandlers {
   public readonly charge: Charge;
 
   /**
-   * @constructor
-   * @TODO: provide manager
+   * @public
    */
-  private constructor() {
-    this.customer = new Customer();
-    this.charge = new Charge();
+  public readonly applicationFee: ApplicationFee;
+
+  /**
+   * @public
+   */
+  public readonly paymentIntent: PaymentIntent;
+
+  /**
+   * @public
+   */
+  public readonly transfer: Transfer;
+
+  /**
+   * @public
+   */
+  public readonly paymentMethod: PaymentMethod;
+
+  /**
+   * @public
+   */
+  public readonly setupIntent: SetupIntent;
+
+  /**
+   * @public
+   */
+  public readonly account: Account;
+
+  /**
+   * @constructor
+   */
+  private constructor(manager: EntityManager) {
+    this.customer = new Customer(manager);
+    this.charge = new Charge(manager);
+    this.applicationFee = new ApplicationFee(manager);
+    this.paymentIntent = new PaymentIntent(manager);
+    this.transfer = new Transfer(manager);
+    this.paymentMethod = new PaymentMethod(manager);
+    this.setupIntent = new SetupIntent(manager);
+    this.account = new Account(manager);
   }
 
   /**
    * Init service
    */
-  public static init(): WebhookHandlers {
-    return new WebhookHandlers();
+  public static init(manager: EntityManager): WebhookHandlers {
+    return new WebhookHandlers(manager);
   }
 }
